customer-feedback-rag/requirements.txt
python-dotenv
openai
flask
requests
numpy
pandas
faker
supabase